Move tests to tensor.test.js
